Move EmbedFileHandler cog to its own module
